requirements.txt
# Neo4j driver
neo4j>=5.14.0

# Vector database - ChromaDB
chromadb>=0.4.22

# Embedding models
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# API Server
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
requests>=2.28.0

# Agent Integration
langchain>=0.0.267
openai>=1.0.0

# Testing
pytest>=7.4.0